refactor(view): move overlay rendering out of view.go
Every modal shares one shape: build lines, wrap them in a frame, place the
frame over the dashboard. Collecting them makes that shape visible instead of
repeating it between panel code.

Declarations move unchanged.
